run-cache: don't forget to account for params
Fixes #3827
