+++ tests/unit/stage/test_stage.py
@@ -102,55 +102,3 @@ def test_always_changed(dvc):
     with dvc.lock:
         assert stage.changed()
         assert stage.status()["path"] == ["always changed"]
-
-
-def test_stage_cache(tmp_dir, dvc, run_copy, mocker):
-    tmp_dir.gen("dep", "dep")
-    tmp_dir.gen(
-        "script.py",
-        (
-            'open("out", "w+").write("out"); '
-            'open("out_no_cache", "w+").write("out_no_cache")'
-        ),
-    )
-    stage = dvc.run(
-        cmd="python script.py",
-        deps=["script.py", "dep"],
-        outs=["out"],
-        outs_no_cache=["out_no_cache"],
-        single_stage=True,
-    )
-
-    with dvc.lock, dvc.state:
-        stage.remove(remove_outs=True, force=True)
-
-    assert not (tmp_dir / "out").exists()
-    assert not (tmp_dir / "out_no_cache").exists()
-    assert not (tmp_dir / "out.dvc").exists()
-
-    cache_dir = os.path.join(
-        dvc.stage_cache.cache_dir,
-        "10",
-        "10b45372fdf4ec14d3f779c5b256378d7a12780e4c7f549a44138e492f098bfe",
-    )
-    cache_file = os.path.join(
-        cache_dir,
-        "bb32e04c6da96a7192513390acedbe4cd6123f8fe5b0ba5fffe39716fe87f6f4",
-    )
-
-    assert os.path.isdir(cache_dir)
-    assert os.listdir(cache_dir) == [os.path.basename(cache_file)]
-    assert os.path.isfile(cache_file)
-
-    run_spy = mocker.patch("dvc.stage.run.cmd_run")
-    checkout_spy = mocker.spy(stage, "checkout")
-    with dvc.lock, dvc.state:
-        stage.run()
-
-    assert not run_spy.called
-    assert checkout_spy.call_count == 1
-
-    assert (tmp_dir / "out").exists()
-    assert (tmp_dir / "out_no_cache").exists()
-    assert (tmp_dir / "out").read_text() == "out"
-    assert (tmp_dir / "out_no_cache").read_text() == "out_no_cache"
